allow synbot6 to respond to >1 item per utterance
It does this by setting up a grammar, parsing out the interesting bits, and responding to them in action methods.

The code that this replaces not only latches on to the first thing it sees, it also *returns early* (not considering
later things) if the first thing turned out to be uninteresting. This version does not do that.

After some consideration, I decided that the grammar rules should do as little as possible. Just the parsing.
The actions then get to decide whether the parsed-out item deserves a reaction of some sort.

Yes, the actions class is nested inside of the `said` method. This is the easiest way for it
to get privileged access to the method parameter `$e`.